Move accessibiltiy help out of terminal instance
Depends on #175785
