+++ src/vs/workbench/contrib/terminal/browser/terminal.contribution.ts
@@ -290,4 +290,5 @@ setupTerminalMenus();
 
 registerColors();
 
+import './accessibility/terminal.accessibility.contribution';
 import './find/terminal.find.contribution';
