+++ src/vs/workbench/contrib/terminal/browser/terminal.ts
@@ -218,8 +218,6 @@ export interface ITerminalService extends ITerminalInstanceHost {
 
 	getEditingTerminal(): ITerminalInstance | undefined;
 	setEditingTerminal(instance: ITerminalInstance | undefined): void;
-
-	showTerminalAccessibilityHelp(): void;
 }
 export class TerminalLinkQuickPickEvent extends MouseEvent {
 
@@ -929,11 +927,6 @@ export interface ITerminalInstance {
 	freePortKillProcess(port: string, commandToRun: string): Promise<void>;
 
 	/**
-	 * Shows the accessibility help widget
-	 */
-	showAccessibilityHelp(): void;
-
-	/**
 	 * Selects the previous suggestion if the suggest widget is visible.
 	 */
 	selectPreviousSuggestion(): void;
@@ -984,8 +977,8 @@ export interface ITerminalInstance {
 // future these should probably be consolidated.
 export interface ITerminalChildElement {
 	element: HTMLElement;
-	layout(dimension: Dimension): void;
-	xtermReady(xterm: IXtermTerminal): void;
+	layout?(dimension: Dimension): void;
+	xtermReady?(xterm: IXtermTerminal): void;
 }
 
 export interface IXtermTerminal {
